refactor(budget): update reducer and and dispatching action when setting a new budget
