+++ src/components/Expenses/SetBudgetModal.jsx
@@ -6,7 +6,6 @@ import {
   DialogTitle,
   TextField,
 } from "@mui/material";
-import useAxiosPrivate from "hooks/useAxiosPrivate";
 import { Context } from "provider/Provider";
 import { DatePicker, LocalizationProvider } from "@mui/x-date-pickers";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
@@ -14,8 +13,7 @@ import Swal from "sweetalert2";
 
 export default function SetBudgetModal() {
   const [open, setOpen] = useState(false);
-  const axiosPrivate = useAxiosPrivate();
-  const { authToken } = useContext(Context);
+  const { createBudget } = useContext(Context);
   const [limit, setLimit] = useState("");
   const [month, setMonth] = useState("");
 
@@ -32,24 +30,12 @@ export default function SetBudgetModal() {
     }
 
     try {
-      const headers = {
-        Authorization: `Bearer ${authToken}`,
-        "Content-Type": "application/json",
-      };
-
       const data = {
         limit,
         month: month.format("YYYY-MM-DD"),
       };
 
-      console.log(data);
-
-      const response = await axiosPrivate.post(
-        `${process.env.REACT_APP_BACKEND_API}budget/`,
-        data,
-        { headers }
-      );
-
+      const response = await createBudget(data);
       if (response.status === 201) {
         setOpen(false);
         setLimit("");
@@ -57,14 +43,12 @@ export default function SetBudgetModal() {
 
         Swal.fire({
           icon: "success",
-          title: "Budget set successfully",
-          timer: 2000,
+          title: "Successfully set the budget",
+          timer: 2300,
           timerProgressBar: true,
         });
       }
-    } catch (error) {
-      console.log(error);
-    }
+    } catch (error) {}
   };
 
   return (
